JBRULES-708 MVEL Integration -broke test to make it work
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@12124 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ kogito-drools/drools-compiler/src/test/java/org/drools/integrationtests/MVELTest.java
@@ -37,7 +37,7 @@ public class MVELTest extends TestCase {
 
         workingMemory.assertObject( new Cheese("stilton", 10) );
         workingMemory.fireAllRules();
-        assertEquals( new Integer(30), list.get(0));
+        assertEquals( new Integer(1020), list.get(0));
     }
     
     public Object compiledExecute(String ex) {
